+++ ft_putchar_fd_rtnint.c
@@ -6,13 +6,13 @@
 /*   By: sabamikan <sabamikan@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/19 23:46:12 by sabamikan         #+#    #+#             */
-/*   Updated: 2023/02/20 14:20:18 by sabamikan        ###   ########.fr       */
+/*   Updated: 2023/03/07 15:15:34 by sabamikan        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_putchar_fd_rtnint(char c, int fd)
+size_t	ft_putchar_fd_rtnint(char c, int fd)
 {
-	return (write(fd, &c, 1));
+	return ((size_t)write(fd, &c, 1));
 }
